♻️ refactor: move auth middleware
